Kill all dashboards and move the one function that was being used in that folder out to utils
