--- autumn/dashboards/multicountry_plots/__init__.py
@@ -1,1 +1,0 @@
-from .run import run_dashboard
