--- autumn/dashboards/philippines/__init__.py
@@ -1,1 +1,0 @@
-from .run import run_dashboard
