+++ docs/papers/covid_19/preprocess/mixing_matrix/mixing_and_npis_vic2021.tex
@@ -1,0 +1,65 @@
+\section{Implementation of non-pharmaceutical interventions} \label{npi}
+A major part of the rationale for the development of this model was to capture the past impact of non-pharmaceutical interventions (NPIs) and produce future scenarios projections with the implementation or release of such interventions.
+
+\subsection{Isolation and quarantine}
+For persons who are identified with symptomatic disease and enter clinical stratum 3, self-isolation is assumed to occur and their infectiousness is modified as described above. The proportion of ambulatory symptomatic persons passively identified through the public health response is determined by the case detection rate as described above in Section \ref{cdr}.
+
+\subsection{Community quarantine or ``lockdown" measures}
+For all NPIs relating to reduction of human mobility or “lockdown” (i.e. all NPIs other than isolation and quarantine), these interventions are implemented through dynamic adjustments to the age-assortative mixing matrix. The baseline mixing matrices have the major advantage of allowing for disaggregation of total contact rates by location, i.e. home, work, school and other locations. This disaggregation allows for the simulation of various NPIs in the local context by dynamically varying the contribution of each location to reflect the historical implementation of the interventions.
+
+For each location $L$ (home, school, work, other locations) the age-specific contact matrix $\mathbf{C^L} = (c_{i,j}^L) \in \mathbb{R}_{+}^{16 \times 16}$ is defined such that $c_{i,j}^L$ is the average number of contacts that a typical individual aged $i$ has with individuals aged $j$. The original matrices from the United Kingdom are denoted $\mathbf{Q^L} = (q_{i,j}^L) \in \mathbb{R}_{+}^{16 \times 16}$, where $q_{i,j}^L$ is defined using the same convention as for $c_{i,j}^L$. The matrices $\mathbf{Q^L}$ were extracted using the R package ``socialmixr'' (v 0.1.8).
+
+Let $\pi_j$ denote the proportion of people aged $j$ in the UK, and $\rho_j$ the proportion of people aged $j$ in Victoria. The contact matrices $\mathbf{C^L}$ were obtained from:
+$$
+c_{i,j}^L = q_{i,j}^L \times \frac{\pi_j}{\rho_j} . 
+$$
+
+
+The overall contact matrix results from the summation of the four location-specific contact matrices: \(C_{0}=C_{H}+C_{S}+C_{W}+C_{L}\), where \(C_{H}\), \(C_{S}\), \(C_{W}\) and \(C_{L}\) are the age-specific contact matrices associated with households, schools, workplaces and other locations, respectively.
+
+In our model, the contributions of the matrices \(C_{S}\), \(C_{W}\) and \(C_{L}\) vary with time such that the input contact matrix can be written:
+\[C(t)= C_{H}+ s(t)^{2}C_{S}+ w(t)^{2}C_{W}+l(t)^{2}C_{L}\]
+
+The modifying functions are each squared to capture the effect of the mobility changes on both the infector and the infectee in any given interaction that could potentially result in transmission. The modifying functions incorporate both macro-distancing and microdistancing effects, depending on the location.
+
+\subsection{School closures/re-openings}
+Reduced attendance at schools is represented through the function \(s(t)\), which represents the proportion of all students currently attending on-site teaching. If schools are fully closed, \(s(t)=0\) and \(C_{S}\) does not contribute to the overall mixing matrix \(C(t)\). \(s(t)\) is calculated through a series of estimates of the proportion of students attending schools, to which a smoothed step function is fitted. Note that the dramatic changes in this contribution to the mixing matrix with school closures/re-openings is a more marked change than is seen with the simulation of policy changes in workplaces and other locations (which are determined by empiric data and so do not vary so abruptly or reach a value of zero).
+
+\subsection{Workplace closures}
+Workplace closures are represented by quadratically reducing the contribution of workplace contacts to the total mixing matrix over time. This is achieved through the scaling term \(w(t)^{2}\) which modifies the contribution of \(C_{W}\) to the overall mixing matrix \(C(t)\). The profile of the function \(w(t)\) is set by fitting a polynomial spline function to Google mobility data for workplace attendance (Table \ref{tab:mobility_map}).
+
+\subsection{Community-wide movement restriction}
+Community-wide movement restriction (or ``lockdown") measures are represented by proportionally reducing the contribution of the other locations contacts to the total mixing matrix over time. This is achieved through the scaling term \(l(t)^{2}\) which modifies the contribution of \(C_{L}\) to the overall mixing matrix \(C(t)\). The profile of the function \(l(t)\) is set by fitting a polynomial spline function to an average of Google mobility data for various locations, as indicated in Table \ref{tab:mobility_map}.
+
+\subsection{Household contacts}
+The contribution of household contacts to the overall mixing matrix \(C(t)\) is fixed over time. Although Google provides mobility estimates for residential contacts, the nature of these data are different from those for each of the other Google mobility types. They represent the time spent in that location, as opposed to the other categories, which measure a change in total visitors rather than the duration. The daily frequency with which people attend their residence is likely to be close to one and we considered that household members likely have a daily opportunity for infection with each other household member. Therefore, we did not implement a function to scale the contribution of household contacts to the mixing matrix with time.
+
+\begin{table}[ht]
+\renewcommand{\baselinestretch}{1}
+    \begin{tabular}{| p{4.4cm} | p{4.4cm} | p{5cm} |}
+        \hline
+        \textbf{location} & \textbf{Approach} & \textbf{Google mobility types} \\
+        \hline
+        School & Policy response & Not applicable \\
+      \hline
+      Household & Constant & Not applicable \\
+      \hline
+      Workplace & Google mobility & Workplace \\
+      \hline
+      Other locations & Google mobility & 
+      Unweighted average of: \begin{itemize}
+			\item Retail and recreation
+          \item Grocery and pharmacy
+          \item Parks
+          \item Transit stations
+      \end{itemize}\\
+      \hline
+    \end{tabular}
+    \title{Mapping of Google mobility data to contact locations.}
+    \caption{\textbf{Mapping of Google mobility data to contact locations.}}
+    \label{tab:mobility_map}
+\end{table}
+
+\subsection{Microdistancing}
+\label{microdist}
+Interventions other than those that prevent people coming into contact with one another are thought to be important to COVID-19 transmission and epidemiology, such as maintaining interpersonal physical distance and the wearing of face coverings. We therefore implemented a ``microdistancing” function to represent reductions in the rate of effective contact that is not attributable to persons visiting specific locations and so is not captured through Google mobility data. This microdistancing function reduces the values of all non-household contributions to the mixing matrices by a certain proportion. These time-varying functions multiplicatively scale the location-specific contact rate modifiers \(s(t)\), \(w(t)\) and \(l(t)\).
